refactor(api): move user anonymized event logging job repository
in src/identity-access-management
